requirements.txt
asgiref==3.7.2
cffi @ file:///private/tmp/cffi-20231211-40239-j25k6o/cffi-1.16.0
coverage==7.3.3
Django==5.0
djangorestframework==3.14.0
docutils @ file:///private/tmp/docutils-20231211-36728-9mhbst/docutils-0.20.1
protobuf==4.21.12
psycopg==3.1.14
psycopg2==2.9.9
pycparser @ file:///private/tmp/pycparser-20231211-25057-cypg40/pycparser-2.21
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
